[FIX] : article-refactor / #55
티켓 수령하기 post api 호출 안되는 오류 해결

+++ app/src/main/java/com/indipage/data/api/ArticleDetailApiService.kt
@@ -22,7 +22,7 @@ interface ArticleDetailApiService {
 
     @POST("/user/ticket/{spaceId}")
     suspend fun postTicketReceive(
-        @Path(value = "spaceId") spaceId: Long
+        @Path(value = "spaceId") spaceId: Int
     ): QRResponse
 
 }
